cleanup and simplify code for only new style clustering
