+++ Core/src/org/sleuthkit/autopsy/timeline/ui/detailview/EventStripeNode.java
@@ -7,30 +7,152 @@ package org.sleuthkit.autopsy.timeline.ui.detailview;
 
 import com.google.common.collect.Range;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import static java.util.Objects.nonNull;
+import java.util.Set;
+import java.util.concurrent.ExecutionException;
+import java.util.logging.Level;
 import java.util.stream.Collectors;
+import javafx.application.Platform;
+import javafx.beans.property.SimpleObjectProperty;
+import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
+import javafx.scene.Cursor;
+import javafx.scene.Node;
+import javafx.scene.control.Button;
+import javafx.scene.control.ContextMenu;
+import javafx.scene.control.Label;
+import javafx.scene.control.OverrunStyle;
+import javafx.scene.control.SeparatorMenuItem;
+import javafx.scene.control.Tooltip;
+import javafx.scene.effect.DropShadow;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.scene.input.MouseButton;
+import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.Background;
 import javafx.scene.layout.BackgroundFill;
+import javafx.scene.layout.Border;
+import javafx.scene.layout.BorderStroke;
+import javafx.scene.layout.BorderStrokeStyle;
+import javafx.scene.layout.BorderWidths;
+import javafx.scene.layout.CornerRadii;
 import javafx.scene.layout.HBox;
+import javafx.scene.layout.Pane;
+import javafx.scene.layout.Priority;
 import javafx.scene.layout.Region;
 import static javafx.scene.layout.Region.USE_PREF_SIZE;
+import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
+import javafx.scene.paint.Color;
+import org.apache.commons.lang3.StringUtils;
+import org.controlsfx.control.action.Action;
+import org.controlsfx.control.action.ActionUtils;
+import org.joda.time.DateTime;
+import org.joda.time.Interval;
+import org.openide.util.NbBundle;
 import org.sleuthkit.autopsy.coreutils.ColorUtilities;
+import org.sleuthkit.autopsy.coreutils.LoggedTask;
+import org.sleuthkit.autopsy.coreutils.Logger;
+import org.sleuthkit.autopsy.timeline.TimeLineController;
 import org.sleuthkit.autopsy.timeline.datamodel.EventCluster;
 import org.sleuthkit.autopsy.timeline.datamodel.EventStripe;
-import static org.sleuthkit.autopsy.timeline.ui.detailview.AbstractDetailViewNode.show;
+import org.sleuthkit.autopsy.timeline.datamodel.FilteredEventsModel;
+import org.sleuthkit.autopsy.timeline.datamodel.TimeLineEvent;
+import org.sleuthkit.autopsy.timeline.datamodel.eventtype.EventType;
+import org.sleuthkit.autopsy.timeline.filters.DescriptionFilter;
+import org.sleuthkit.autopsy.timeline.filters.RootFilter;
+import org.sleuthkit.autopsy.timeline.filters.TypeFilter;
+import org.sleuthkit.autopsy.timeline.zooming.DescriptionLOD;
+import org.sleuthkit.autopsy.timeline.zooming.EventTypeZoomLevel;
+import org.sleuthkit.autopsy.timeline.zooming.ZoomParams;
+import org.sleuthkit.datamodel.SleuthkitCase;
+import org.sleuthkit.datamodel.TskCoreException;
 
 /**
  *
  */
-public class EventStripeNode extends AbstractDetailViewNode<EventStripe, EventStripeNode> {
+final public class EventStripeNode extends StackPane {
 
     private final HBox rangesHBox = new HBox();
+    private final EventStripe eventStripe;
+    private Tooltip tooltip;
 
-    EventStripeNode(EventStripe eventStripe, EventStripeNode parentNode, EventDetailChart chart) {
-        super(chart, eventStripe, parentNode);
+    public EventStripeNode(EventDetailChart chart, EventStripe eventStripe, EventStripeNode parentEventNode) {
+        this.eventStripe = eventStripe;
+        this.parentNode = parentEventNode;
+        this.chart = chart;
+        descLOD.set(eventStripe.getDescriptionLOD());
+        sleuthkitCase = chart.getController().getAutopsyCase().getSleuthkitCase();
+        eventsModel = chart.getController().getEventsModel();
+        ImageView hashIV = new ImageView(HASH_PIN);
+        ImageView tagIV = new ImageView(TAG);
+        if (eventStripe.getEventIDsWithHashHits().isEmpty()) {
+            show(hashIV, false);
+        }
+        if (eventStripe.getEventIDsWithTags().isEmpty()) {
+            show(tagIV, false);
+        }
+
+        expandClusterAction = new ExpandClusterAction();
+        plusButton = ActionUtils.createButton(expandClusterAction, ActionUtils.ActionTextBehavior.HIDE);
+        configureLODButton(plusButton);
+
+        collapseClusterAction = new CollapseClusterAction();
+        minusButton = ActionUtils.createButton(collapseClusterAction, ActionUtils.ActionTextBehavior.HIDE);
+        configureLODButton(minusButton);
+
+        HBox.setHgrow(spacer, Priority.ALWAYS);
+        header = new HBox(getDescrLabel(), getCountLabel(), hashIV, tagIV, minusButton, plusButton);
+
+        header.setMinWidth(USE_PREF_SIZE);
+        header.setPadding(new Insets(2, 5, 2, 5));
+        header.setAlignment(Pos.CENTER_LEFT);
+        //setup description label
+        evtColor = getEventType().getColor();
+
+        eventTypeImageView.setImage(getEventType().getFXImage());
+        descrLabel.setGraphic(eventTypeImageView);
+        descrLabel.setPrefWidth(USE_COMPUTED_SIZE);
+        descrLabel.setTextOverrun(OverrunStyle.CENTER_ELLIPSIS);
+        descrLabel.setMouseTransparent(true);
+
+        //set up subnode pane sizing contraints
+        subNodePane.setPrefHeight(USE_COMPUTED_SIZE);
+        subNodePane.setMinHeight(USE_PREF_SIZE);
+        subNodePane.setMinWidth(USE_PREF_SIZE);
+        subNodePane.setMaxHeight(USE_PREF_SIZE);
+        subNodePane.setMaxWidth(USE_PREF_SIZE);
+        subNodePane.setPickOnBounds(false);
+
+        setAlignment(Pos.TOP_LEFT);
+        setMinHeight(24);
+        setPrefHeight(USE_COMPUTED_SIZE);
+        setMaxHeight(USE_PREF_SIZE);
+        setOnMouseClicked(new EventMouseHandler());
+
+        //set up mouse hover effect and tooltip
+        setOnMouseEntered((MouseEvent e) -> {
+            //defer tooltip creation till needed, this had a surprisingly large impact on speed of loading the chart
+            installTooltip();
+            showDescriptionLoDControls(true);
+            toFront();
+        });
+
+        setOnMouseExited((MouseEvent e) -> {
+            showDescriptionLoDControls(false);
+        });
+        setCursor(Cursor.HAND);
+
+        setBackground(new Background(new BackgroundFill(evtColor.deriveColor(0, 1, 1, .1), CORNER_RADII, Insets.EMPTY)));
+
+        setLayoutX(getChart().getXAxis().getDisplayPosition(new DateTime(eventStripe.getStartMillis())) - getLayoutXCompensation());
+
         minWidthProperty().bind(rangesHBox.widthProperty());
         final VBox internalVBox = new VBox(header, getSubNodePane());
         internalVBox.setAlignment(Pos.CENTER_LEFT);
@@ -51,13 +173,15 @@ public class EventStripeNode extends AbstractDetailViewNode<EventStripe, EventSt
      *
      * @param showControls the value of par
      */
-    @Override
     void showDescriptionLoDControls(final boolean showControls) {
-        super.showDescriptionLoDControls(showControls);
+        DropShadow dropShadow = dropShadowMap.computeIfAbsent(getEventType(),
+                eventType -> new DropShadow(10, eventType.getColor()));
+        getSpanFillNode().setEffect(showControls ? dropShadow : null);
+        show(minusButton, showControls);
+        show(plusButton, showControls);
         show(getSpacer(), showControls);
     }
 
-    @Override
     public void setSpanWidths(List<Double> spanWidths) {
         for (int i = 0; i < spanWidths.size(); i++) {
             Region spanRegion = (Region) rangesHBox.getChildren().get(i);
@@ -69,40 +193,450 @@ public class EventStripeNode extends AbstractDetailViewNode<EventStripe, EventSt
     }
 
     EventStripe getStripe() {
-        return getEventBundle();
+        return eventStripe;
     }
 
-    @Override
     HBox getSpanFillNode() {
         return rangesHBox;
-    }
-
-    @Override
-    Collection<EventStripe> makeBundlesFromClusters(List<EventCluster> eventClusters) {
-        return eventClusters.stream().collect(
-                Collectors.toMap(
-                        EventCluster::getDescription, //key
-                        EventStripe::new, //value
-                        EventStripe::merge)//merge method
-        ).values();
     }
 
     /**
      *
      * @param showSpans the value of showSpans
      */
-    @Override
     void showSpans(final boolean showSpans) {
         rangesHBox.setVisible(showSpans);
     }
 
-    @Override
-    void installTooltip() {
-//        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    synchronized void installTooltip() {
+        //TODO: all this work should probably go on a background thread...
+        if (tooltip == null) {
+            HashMap<String, Long> hashSetCounts = new HashMap<>();
+            if (!getStripe().getEventIDsWithHashHits().isEmpty()) {
+                hashSetCounts = new HashMap<>();
+                try {
+                    for (TimeLineEvent tle : getEventsModel().getEventsById(getStripe().getEventIDsWithHashHits())) {
+                        Set<String> hashSetNames = getSleuthkitCase().getAbstractFileById(tle.getFileID()).getHashSetNames();
+                        for (String hashSetName : hashSetNames) {
+                            hashSetCounts.merge(hashSetName, 1L, Long::sum);
+                        }
+                    }
+                } catch (TskCoreException ex) {
+                    LOGGER.log(Level.SEVERE, "Error getting hashset hit info for event.", ex);
+                }
+            }
+
+            Map<String, Long> tagCounts = new HashMap<>();
+            if (getEventStripe().getEventIDsWithTags().isEmpty() == false) {
+                tagCounts.putAll(getEventsModel().getTagCountsByTagName(getEventStripe().getEventIDsWithTags()));
+            }
+
+            String hashSetCountsString = hashSetCounts.entrySet().stream()
+                    .map((Map.Entry<String, Long> t) -> t.getKey() + " : " + t.getValue())
+                    .collect(Collectors.joining("\n"));
+            String tagCountsString = tagCounts.entrySet().stream()
+                    .map((Map.Entry<String, Long> t) -> t.getKey() + " : " + t.getValue())
+                    .collect(Collectors.joining("\n"));
+
+            tooltip = new Tooltip(
+                    NbBundle.getMessage(this.getClass(), "AggregateEventNode.installTooltip.text",
+                            getEventStripe().getEventIDs().size(), getEventStripe().getEventType(), getEventStripe().getDescription(),
+                            getEventStripe().getStart().toString(TimeLineController.getZonedFormatter()),
+                            getEventStripe().getEnd().toString(TimeLineController.getZonedFormatter()))
+                    + (hashSetCountsString.isEmpty() ? "" : "\n\nHash Set Hits\n" + hashSetCountsString)
+                    + (tagCountsString.isEmpty() ? "" : "\n\nTags\n" + tagCountsString)
+            );
+            Tooltip.install(this, tooltip);
+        }
     }
 
-    @Override
     EventStripeNode getNodeForBundle(EventStripe cluster) {
-        return new EventStripeNode(cluster, this, getChart());
+        return new EventStripeNode(getChart(), cluster, this);
+    }
+
+    static final Image HASH_PIN = new Image("/org/sleuthkit/autopsy/images/hashset_hits.png");
+    static final Image PLUS = new Image("/org/sleuthkit/autopsy/timeline/images/plus-button.png"); // NON-NLS
+    static final Image MINUS = new Image("/org/sleuthkit/autopsy/timeline/images/minus-button.png"); // NON-NLS
+    static final Image TAG = new Image("/org/sleuthkit/autopsy/images/green-tag-icon-16.png"); // NON-NLS
+    static final CornerRadii CORNER_RADII = new CornerRadii(3);
+    /**
+     * the border to apply when this node is 'selected'
+     */
+    static final Border selectionBorder = new Border(new BorderStroke(Color.BLACK, BorderStrokeStyle.SOLID, CORNER_RADII, new BorderWidths(2)));
+    private static final Logger LOGGER = Logger.getLogger(EventStripeNode.class
+            .getName());
+
+    static void configureLODButton(Button b) {
+        b.setMinSize(16, 16);
+        b.setMaxSize(16, 16);
+        b.setPrefSize(16, 16);
+        show(b, false);
+    }
+
+    static void show(Node b, boolean show) {
+        b.setVisible(show);
+        b.setManaged(show);
+    }
+    private final Map<EventType, DropShadow> dropShadowMap = new HashMap<>();
+    final Color evtColor;
+
+    private final EventStripeNode parentNode;
+    private DescriptionVisibility descrVis;
+
+    EventType getEventType() {
+        return eventStripe.getEventType();
+    }
+
+    String getDescription() {
+        return eventStripe.getDescription();
+    }
+
+    long getStartMillis() {
+        return eventStripe.getStartMillis();
+    }
+
+    /**
+     * Pane that contains AggregateEventNodes of any 'subevents' if they are
+     * displayed
+     *
+     * //TODO: move more of the control of subnodes/events here and out of
+     * EventDetail Chart
+     */
+    private final Pane subNodePane = new Pane();
+
+    Pane getSubNodePane() {
+        return subNodePane;
+    }
+
+    /**
+     * The ImageView used to show the icon for this node's event's type
+     */
+    private final ImageView eventTypeImageView = new ImageView();
+
+    /**
+     * The label used to display this node's event's description
+     */
+    final Label descrLabel = new Label();
+
+    /**
+     * The label used to display this node's event count
+     */
+    final Label countLabel = new Label();
+
+    private final EventDetailChart chart;
+    private final SleuthkitCase sleuthkitCase;
+
+    SleuthkitCase getSleuthkitCase() {
+        return sleuthkitCase;
+    }
+
+    FilteredEventsModel getEventsModel() {
+        return eventsModel;
+    }
+    private final FilteredEventsModel eventsModel;
+
+    private final Button plusButton;
+    private final Button minusButton;
+
+    private final SimpleObjectProperty<DescriptionLOD> descLOD = new SimpleObjectProperty<>();
+    final HBox header;
+
+    Region getSpacer() {
+        return spacer;
+    }
+
+    private final Region spacer = new Region();
+
+    private final CollapseClusterAction collapseClusterAction;
+    private final ExpandClusterAction expandClusterAction;
+
+    @SuppressWarnings("unchecked")
+    public List<EventStripeNode> getSubNodes() {
+        return subNodePane.getChildrenUnmodifiable().stream()
+                .map(t -> (EventStripeNode) t)
+                .collect(Collectors.toList());
+    }
+
+    /**
+     * apply the 'effect' to visually indicate selection
+     *
+     * @param applied true to apply the selection 'effect', false to remove it
+     */
+    public void applySelectionEffect(boolean applied) {
+        Platform.runLater(() -> {
+            if (applied) {
+                setBorder(selectionBorder);
+            } else {
+                setBorder(null);
+            }
+        });
+    }
+
+    /**
+     * make a new filter intersecting the global filter with description and
+     * type filters to restrict sub-clusters
+     *
+     */
+    RootFilter getSubClusterFilter() {
+        RootFilter subClusterFilter = eventsModel.filterProperty().get().copyOf();
+        subClusterFilter.getSubFilters().addAll(
+                new DescriptionFilter(eventStripe.getDescriptionLOD(), eventStripe.getDescription()),
+                new TypeFilter(getEventType()));
+        return subClusterFilter;
+    }
+
+    /**
+     * @param w the maximum width the description label should have
+     */
+    public void setDescriptionWidth(double w) {
+        getDescrLabel().setMaxWidth(w);
+    }
+
+    /**
+     * apply the 'effect' to visually indicate highlighted nodes
+     *
+     * @param applied true to apply the highlight 'effect', false to remove it
+     */
+    public synchronized void applyHighlightEffect(boolean applied) {
+        if (applied) {
+            getDescrLabel().setStyle("-fx-font-weight: bold;"); // NON-NLS
+            getSpanFillNode().setBackground(new Background(new BackgroundFill(getEventType().getColor().deriveColor(0, 1, 1, .3), CORNER_RADII, Insets.EMPTY)));
+            setBackground(new Background(new BackgroundFill(getEventType().getColor().deriveColor(0, 1, 1, .2), CORNER_RADII, Insets.EMPTY)));
+        } else {
+            getDescrLabel().setStyle("-fx-font-weight: normal;"); // NON-NLS
+            getSpanFillNode().setBackground(new Background(new BackgroundFill(getEventType().getColor().deriveColor(0, 1, 1, .1), CORNER_RADII, Insets.EMPTY)));
+            setBackground(new Background(new BackgroundFill(getEventType().getColor().deriveColor(0, 1, 1, .1), CORNER_RADII, Insets.EMPTY)));
+        }
+    }
+
+    String getDisplayedDescription() {
+        return getDescrLabel().getText();
+    }
+
+    Button getPlusButton() {
+        return plusButton;
+    }
+
+    Button getMinusButton() {
+        return minusButton;
+    }
+
+    public final Label getDescrLabel() {
+        return descrLabel;
+    }
+
+    final public Label getCountLabel() {
+        return countLabel;
+    }
+
+    public EventStripeNode getParentNode() {
+        return parentNode;
+    }
+
+    public final EventDetailChart getChart() {
+        return chart;
+    }
+
+    public DescriptionLOD getDescLOD() {
+        return descLOD.get();
+    }
+
+    /**
+     * loads sub-bundles at the given Description LOD, continues
+     *
+     * @param requestedDescrLoD
+     * @param expand
+     */
+    private synchronized void loadSubBundles(DescriptionLOD.RelativeDetail relativeDetail) {
+        subNodePane.getChildren().clear();
+        if (descLOD.get().withRelativeDetail(relativeDetail) == eventStripe.getDescriptionLOD()) {
+            descLOD.set(eventStripe.getDescriptionLOD());
+            showSpans(true);
+            chart.setRequiresLayout(true);
+            chart.requestChartLayout();
+        } else {
+            showSpans(false);
+
+            // make new ZoomParams to query with
+            final RootFilter subClusterFilter = getSubClusterFilter();
+            /*
+             * We need to extend end time because for the query by one second,
+             * because it is treated as an open interval but we want to include
+             * events at exactly the time of the last event in this cluster
+             */
+            final Interval subClusterSpan = new Interval(eventStripe.getStartMillis(), eventStripe.getEndMillis() + 1000);
+            final EventTypeZoomLevel eventTypeZoomLevel = eventsModel.eventTypeZoomProperty().get();
+            final ZoomParams zoomParams = new ZoomParams(subClusterSpan, eventTypeZoomLevel, subClusterFilter, getDescLOD());
+
+            LoggedTask<Set<EventStripeNode>> loggedTask = new LoggedTask<Set<EventStripeNode>>(
+                    NbBundle.getMessage(this.getClass(), "AggregateEventNode.loggedTask.name"), true) {
+                        private Collection<EventStripe> bundles;
+                        private volatile DescriptionLOD loadedDescriptionLoD = getDescLOD().withRelativeDetail(relativeDetail);
+                        private DescriptionLOD next = loadedDescriptionLoD;
+
+                        @Override
+                        protected Set<EventStripeNode> call() throws Exception {
+                            do {
+                                loadedDescriptionLoD = next;
+                                if (loadedDescriptionLoD == eventStripe.getDescriptionLOD()) {
+                                    return Collections.emptySet();
+                                }
+                                bundles = eventsModel.getEventClusters(zoomParams.withDescrLOD(loadedDescriptionLoD)).stream()
+                                .collect(Collectors.toMap(
+                                                EventCluster::getDescription, //key
+                                                EventStripe::new, //value
+                                                EventStripe::merge) //merge method
+                                ).values();
+                                next = loadedDescriptionLoD.withRelativeDetail(relativeDetail);
+                            } while (bundles.size() == 1 && nonNull(next));
+
+                            // return list of AbstractEventStripeNodes representing sub-bundles
+                            return bundles.stream()
+                            .map(EventStripeNode.this::getNodeForBundle)
+                            .collect(Collectors.toSet());
+                        }
+
+                        @Override
+                        protected void succeeded() {
+                            chart.setCursor(Cursor.WAIT);
+                            try {
+                                Set<EventStripeNode> subBundleNodes = get();
+                                if (subBundleNodes.isEmpty()) {
+                                    showSpans(true);
+                                } else {
+                                    showSpans(false);
+                                }
+                                descLOD.set(loadedDescriptionLoD);
+                                //assign subNodes and request chart layout
+                                subNodePane.getChildren().setAll(subBundleNodes);
+                                chart.setRequiresLayout(true);
+                                chart.requestChartLayout();
+                            } catch (InterruptedException | ExecutionException ex) {
+                                LOGGER.log(Level.SEVERE, "Error loading subnodes", ex);
+                            }
+                            chart.setCursor(null);
+                        }
+                    };
+
+            //start task
+            chart.getController().monitorTask(loggedTask);
+        }
+    }
+
+    double getLayoutXCompensation() {
+        return (getParentNode() != null ? getParentNode().getLayoutXCompensation() : 0)
+                + getBoundsInParent().getMinX();
+    }
+
+    public void setDescriptionVisibility(DescriptionVisibility descrVis) {
+        this.descrVis = descrVis;
+        final int size = eventStripe.getEventIDs().size();
+
+        switch (this.descrVis) {
+            case COUNT_ONLY:
+                descrLabel.setText("");
+                countLabel.setText(String.valueOf(size));
+                break;
+            case HIDDEN:
+                countLabel.setText("");
+                descrLabel.setText("");
+                break;
+            default:
+            case SHOWN:
+                String description = eventStripe.getDescription();
+                description = getParentNode() != null
+                        ? "    ..." + StringUtils.substringAfter(description, getParentNode().getDescription())
+                        : description;
+                descrLabel.setText(description);
+                countLabel.setText(((size == 1) ? "" : " (" + size + ")")); // NON-NLS
+                break;
+        }
+    }
+
+    public EventStripe getEventStripe() {
+        return eventStripe;
+    }
+
+    public Set<Long> getEventsIDs() {
+        return eventStripe.getEventIDs();
+    }
+
+    /**
+     * event handler used for mouse events on {@link AggregateEventNode}s
+     */
+    private class EventMouseHandler implements EventHandler<MouseEvent> {
+
+        private ContextMenu contextMenu;
+
+        @Override
+        public void handle(MouseEvent t) {
+
+            if (t.getButton() == MouseButton.PRIMARY) {
+                t.consume();
+                if (t.isShiftDown()) {
+                    if (chart.selectedNodes.contains(EventStripeNode.this) == false) {
+                        chart.selectedNodes.add(EventStripeNode.this);
+                    }
+                } else if (t.isShortcutDown()) {
+                    chart.selectedNodes.removeAll(EventStripeNode.this);
+                } else if (t.getClickCount() > 1) {
+                    final DescriptionLOD next = descLOD.get().moreDetailed();
+                    if (next != null) {
+                        loadSubBundles(DescriptionLOD.RelativeDetail.MORE);
+
+                    }
+                } else {
+                    chart.selectedNodes.setAll(EventStripeNode.this);
+                }
+                t.consume();
+            } else if (t.getButton() == MouseButton.SECONDARY) {
+                ContextMenu chartContextMenu = chart.getChartContextMenu(t);
+                if (contextMenu == null) {
+                    contextMenu = new ContextMenu();
+                    contextMenu.setAutoHide(true);
+
+                    contextMenu.getItems().add(ActionUtils.createMenuItem(expandClusterAction));
+                    contextMenu.getItems().add(ActionUtils.createMenuItem(collapseClusterAction));
+
+                    contextMenu.getItems().add(new SeparatorMenuItem());
+                    contextMenu.getItems().addAll(chartContextMenu.getItems());
+                }
+                contextMenu.show(EventStripeNode.this, t.getScreenX(), t.getScreenY());
+                t.consume();
+            }
+        }
+    }
+
+    private class ExpandClusterAction extends Action {
+
+        ExpandClusterAction() {
+            super("Expand");
+
+            setGraphic(new ImageView(PLUS));
+            setEventHandler((ActionEvent t) -> {
+                final DescriptionLOD next = descLOD.get().moreDetailed();
+                if (next != null) {
+                    loadSubBundles(DescriptionLOD.RelativeDetail.MORE);
+
+                }
+            });
+            disabledProperty().bind(descLOD.isEqualTo(DescriptionLOD.FULL));
+        }
+    }
+
+    private class CollapseClusterAction extends Action {
+
+        CollapseClusterAction() {
+            super("Collapse");
+
+            setGraphic(new ImageView(MINUS));
+            setEventHandler((ActionEvent t) -> {
+                final DescriptionLOD previous = descLOD.get().lessDetailed();
+                if (previous != null) {
+                    loadSubBundles(DescriptionLOD.RelativeDetail.LESS);
+                }
+            });
+            disabledProperty().bind(descLOD.isEqualTo(eventStripe.getDescriptionLOD()));
+        }
     }
 }
