+++ Core/src/org/sleuthkit/autopsy/timeline/ui/detailview/tree/NavPanel.java
@@ -43,7 +43,6 @@ import org.sleuthkit.autopsy.timeline.TimeLineController;
 import org.sleuthkit.autopsy.timeline.TimeLineView;
 import org.sleuthkit.autopsy.timeline.datamodel.EventCluster;
 import org.sleuthkit.autopsy.timeline.datamodel.FilteredEventsModel;
-import org.sleuthkit.autopsy.timeline.ui.detailview.DetailViewNode;
 import org.sleuthkit.autopsy.timeline.ui.detailview.DetailViewPane;
 
 /**
@@ -91,8 +90,8 @@ public class NavPanel extends BorderPane implements TimeLineView {
         });
         detailViewPane.getSelectedNodes().addListener((Observable observable) -> {
             eventsTree.getSelectionModel().clearSelection();
-            detailViewPane.getSelectedNodes().forEach((DetailViewNode<?> t) -> {
-                eventsTree.getSelectionModel().select(((NavTreeItem) eventsTree.getRoot()).findTreeItemForEvent(t.getEventBundle()));
+            detailViewPane.getSelectedNodes().forEach(eventStripeNode -> {
+                eventsTree.getSelectionModel().select(((NavTreeItem) eventsTree.getRoot()).findTreeItemForEvent(eventStripeNode.getEventStripe()));
             });
         });
